% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classement.R
\name{bdd_classement}
\alias{bdd_classement}
\title{Obtenir le classement actuel des PJ}
\usage{
bdd_classement(actions)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}
}
\value{
data.frame avec le classement
}
\description{
Obtenir le classement actuel des PJ
}
