% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classement.R
\name{copie_classement}
\alias{copie_classement}
\title{Copie le classement des PJ}
\usage{
copie_classement(actions, user_name, PA, duree = NA)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{user_name}{PJ a l'origine de la copie}

\item{PA}{nombre de PA dépensé}

\item{duree}{(pour admin) détermine manuellement la durée de la recherche}
}
\value{
liste avec les nouvelles lignes et le message à afficher
}
\description{
Copie le classement des PJ
}
