% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{copie_indice}
\alias{copie_indice}
\title{Copier l'indice d'un autre PJ}
\usage{
copie_indice(actions, user_name, cible_name, PA)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{user_name}{PJ a l'origine de la copie}

\item{cible_name}{PJ ciblé par la copie}

\item{PA}{nombre de PA dépensé}
}
\value{
liste avec les nouvelles lignes et le message à afficher
}
\description{
`copie_indice()` permet de faire une copie d'une enquête réalisée par un autre PJ.
}
