% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea.R
\name{creation_bdd_des}
\alias{creation_bdd_des}
\title{Fonction de création du dé d'action}
\usage{
creation_bdd_des(actions, bonus_recherche = 0)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{bonus_recherche}{crée le bonus de recherche automatique}
}
\value{
data.frame avec les valeurs du dé
}
\description{
`creation_bdd_des()` permet de choisir le dé pour les aléas des actions.
}
