% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation.R
\name{creation_server_habemus}
\alias{creation_server_habemus}
\title{Création d'un fichier de serveur pour Habemus Papam}
\usage{
creation_server_habemus(
  df_init_habemus,
  df_init_habemus_indices,
  df_init_habemus_indices_dispo
)
}
\arguments{
\item{df_init_habemus}{data.frame avec les infos initiales}

\item{df_init_habemus_indices}{data.frame avec les textes des indices}

\item{df_init_habemus_indices_dispo}{data.frame avec les indices de départ}
}
\value{
data.frame
}
\description{
Création d'un fichier de serveur pour Habemus Papam
}
