% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dieu.R
\name{creer_server_dieu}
\alias{creer_server_dieu}
\title{Créer le serveur Dieu est mort}
\usage{
creer_server_dieu(server_name, path = "./")
}
\arguments{
\item{server_name}{nom du serveur}

\item{df_init_dieu}{lien vers la base de données initiale (joueurs)}

\item{df_init_dieu_indices}{lien vers la base de données initiale (indices)}

\item{df_init_dieu_indices_dispo}{lien vers la base de données initiale (indices dispo)}
}
\value{
data.frame avec toutes les infos de la murder party
}
\description{
`creer_serveur_dieu()` permet de créer un fichier csv avec les données de serveur pour Dieu est mort.
}
\examples{

# Exemple avec les fichiers par défaut
creer_server_dieu('server_dieu_test')
}
