% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dieu.R
\name{creer_serveur_dieu}
\alias{creer_serveur_dieu}
\title{Créer le serveur Dieu est mort}
\usage{
creer_serveur_dieu(
  server_name,
  df_init_dieu = "data-raw/dieu/init_dieu.csv",
  df_init_dieu_indices = "data-raw/dieu/init_dieu_indices.csv",
  df_init_dieu_indices_dispo = "data-raw/dieu/init_dieu_indices_dispo.csv"
)
}
\arguments{
\item{server_name}{nom du serveur}

\item{df_init_dieu}{lien vers la base de données initiale (joueurs)}

\item{df_init_dieu_indices}{lien vers la base de données initiale (indices)}

\item{df_init_dieu_indices_dispo}{lien vers la base de données initiale (indices dispo)}
}
\value{
data.frame avec toutes les infos de la murder party
}
\description{
`creer_serveur_dieu()` permet de créer un fichier csv avec les données de serveur pour Dieu est mort.
}
\examples{

# Exemple avec les fichiers par défaut
creation_server_dieu('server_dieu_test')
}
