% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{display_duration}
\alias{display_duration}
\title{Fonction pour afficher la durée de manière pertinente}
\usage{
display_duration(diff_seconds)
}
\arguments{
\item{diff_seconds}{}
}
\description{
Fonction pour afficher la durée de manière pertinente
}
