% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{envoi_message}
\alias{envoi_message}
\title{Envoyer un message à un ou des PJ}
\usage{
envoi_message(actions, origine, destinataire, message, PA, delai)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{origine}{origine du message}

\item{destinataire}{destinataire du message}

\item{message}{message à envoyer}

\item{PA}{nombre de PA dépensé}

\item{délai}{détermine le temps pour envoyer le message}
}
\value{
liste avec les nouvelles lignes et le message à afficher
}
\description{
`envoi_message()` permet d'envoyer un message à un (ou d'autres) PJ.
}
