% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{info_indices}
\alias{info_indices}
\title{Charger les indices}
\usage{
info_indices(actions)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}
}
\value{
data.frame avec les indices disponibles
}
\description{
Charger les indices
}
