% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{info_user}
\alias{info_user}
\title{Charger les user/password}
\usage{
info_user(actions)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}
}
\value{
data.frame avec les users et passwords
}
\description{
Charger les user/password
}
