% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{interception_indice}
\alias{interception_indice}
\title{Intercepter l'enquête future d'un autre PJ}
\usage{
interception_indice(actions, user_name, cible_name, PA)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{user_name}{PJ a l'origine de l'interception}

\item{cible_name}{PJ ciblé par l'interception}

\item{PA}{nombre de PA dépensé}
}
\value{
liste avec les nouvelles lignes et le message à afficher
}
\description{
`interception_indice()` permet de préparer une interception d'une prochaine recherche d'indice d'un autre PJ.
}
