% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{lancer_serveur_habemus}
\alias{lancer_serveur_habemus}
\title{Lancer le serveur Habemus Papam}
\usage{
lancer_serveur_habemus(server_name, id_chatgpt = NULL)
}
\arguments{
\item{server_name}{nom du serveur}

\item{id_chatgpt}{(facultatif) credential de chatGPT}
}
\value{
exécute l'interface shiny
}
\description{
`lancer_serveur_habemus()` permet de lancer, à partir d'un fichier de données de serveur, l'interface web pour Habemus Papam.
}
\examples{
# Exemple avec les fichiers par défaut
creation_server_habemus('server_habemus_test')
creer_serveur_habemus("server_habemus_test")
}
