% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{liste_indices}
\alias{liste_indices}
\title{Lister les indices actuels d'un PJ}
\usage{
liste_indices(actions, user_name)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{user_name}{PJ dont on veut la liste des indices actuels}
}
\value{
liste avec les indices
}
\description{
`liste_indices()` liste les indices actuels d'un PJ
}
