% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classement.R
\name{modification_classement}
\alias{modification_classement}
\title{Altérer le classement des entités}
\usage{
modification_classement(actions, user_name, cible, PA, duree = NA)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{user_name}{PJ a l'origine de la copie}

\item{cible}{PJ que l'on veut placer en haut du classement}

\item{PA}{nombre de PA dépensé}

\item{duree}{(pour admin) détermine manuellement la durée de la recherche}
}
\value{
liste avec les nouvelles lignes et le message à afficher
}
\description{
Altérer le classement des entités
}
