% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/murderParty-package.R
\docType{package}
\name{murderParty-package}
\alias{murderParty}
\alias{murderParty-package}
\title{murderParty: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tdelc/murderParty}
  \item Report bugs at \url{https://github.com/tdelc/murderParty/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
