% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea.R
\name{plot_bdd_des}
\alias{plot_bdd_des}
\title{Fonction de graphique pour le dé}
\usage{
plot_bdd_des(valeur_alea)
}
\arguments{
\item{valeur_alea}{data.frame provenant de creation_bdd_des()}
}
\value{
ggplot
}
\description{
`plot_bdd_des()` permet de représenter graphiquement les résultats possible du dé.
}
