% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{prepare_timer}
\alias{prepare_timer}
\title{Préparer les timers}
\usage{
prepare_timer(timer)
}
\arguments{
\item{timer}{timer à modifier}
}
\value{
timer en meilleur format
}
\description{
Préparer les timers
}
\examples{
prepare_timer(now())
}
