% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{recherche_indice}
\alias{recherche_indice}
\title{Rechercher un indice}
\usage{
recherche_indice(
  actions,
  user_name,
  enquete,
  PA,
  resultat = NA,
  duree = NA,
  id_chatgpt = NULL
)
}
\arguments{
\item{actions}{objet réactif contenant la base de données du serveur}

\item{user_name}{PJ a l'origine de la recherche}

\item{enquete}{nom de l'enquête}

\item{PA}{nombre de PA dépensé}

\item{resultat}{(pour admin) détermine manuellement le résultat}

\item{duree}{(pour admin) détermine manuellement la durée de la recherche}

\item{id_chatgpt}{(facultatif) credential de chatGPT}
}
\value{
liste avec les nouvelles lignes et le message à afficher
}
\description{
`recherche_indice()` permet de faire une recherche d'indices parmi les indices dispo.
}
