% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{show_message}
\alias{show_message}
\title{Afficher un message dans l'interface}
\usage{
show_message(title, message)
}
\arguments{
\item{title}{titre du message}

\item{message}{texte du message}
}
\value{
modalBox
}
\description{
Afficher un message dans l'interface
}
