% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alea.R
\name{creation_bdd_des}
\alias{creation_bdd_des}
\title{Fonction de création BDD dés}
\usage{
creation_bdd_des(df, bonus_recherche = 0)
}
\arguments{
\item{bonus_recherche}{}
}
\description{
Fonction de création BDD dés
}
