% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{creer_serveur_habemus}
\alias{creer_serveur_habemus}
\title{Créer le serveur Habemus Papam}
\usage{
creer_serveur_habemus(
  server_name,
  df_init_habemus = "data-raw/init_habemus.csv",
  df_init_habemus_indices = "data-raw/init_habemus_indices.csv",
  df_init_habemus_indices_dispo = "data-raw/init_habemus_indices_dispo.csv"
)
}
\arguments{
\item{server_name}{nom du serveur}

\item{df_init_habemus}{lien vers la base de données initiale (joueurs)}

\item{df_init_habemus_indices}{lien vers la base de données initiale (indices)}

\item{df_init_habemus_indices_dispo}{lien vers la base de données initiale (indices dispo)}
}
\value{
data.frame avec toutes les données de la murder Party
}
\description{
`creer_serveur_habemus()` permet de créer un fichier csv avec les données de serveur pour Habemus Papam.
}
\examples{

# Exemple avec les fichiers par défaut
creation_server_habemus('server_habemus_test')
}
