% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions.R
\name{display_duration}
\alias{display_duration}
\title{Afficher la durée de manière propre}
\usage{
display_duration(diff_seconds)
}
\arguments{
\item{diff_seconds}{durée en secondes}
}
\value{
texte avec la durée
}
\description{
Afficher la durée de manière propre
}
\examples{
display_duration(1234)
}
